Push RMQ Module to Open Source
